Added sub menu for New, Edit and Delete buttons
Save and Cancel buttons added to RecordActionButtons. When clicking the New, Edit or Delete button, the user is presented with the option to cancel or commit the action. Delete functionality currently working. Edit and New still to be implemented.

Also linked to AwesomeFonts to include 'delete' icon in Issued component.