Fixed functionality of 'prev/next record' buttons when fetching a record by job number
Moved currentRecordNumber to global state rather than the RecordActionButtons local state. This allows it to be updated by recordUpper as well when typing in a new record number.
Needed to change how currentRecordNumber is assigned in the recordReducer for the prev/next buttons to continue working correctly.
Still an issue when reaching the very first record.